backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
openai==1.58.1
python-dotenv==1.0.1
supabase==2.11.0
